feat: Enhance support for multi-button controllers and add double play (DP) functionality
- Updated architecture documentation to reflect changes in button mapping and input monitoring.
- Introduced a new roadmap document for double play controller support, detailing necessary architectural changes and UI/UX updates.
- Modified input handling to support u32 button masks instead of u16, accommodating controllers with more than 16 buttons.
- Implemented bulk actions in the dashboard for resetting stats and changing switch models.
- Enhanced input tester UI to visually represent IIDX controller layout with piano key styling.
- Added logging functionality to the application for better debugging and monitoring.
- Updated various components and services to ensure compatibility with new input handling and multi-controller support.

+++ src-tauri/src/commands.rs
@@ -34,7 +34,7 @@ pub fn reset_to_default_mapping(state: State<'_, AppState>) {
 }
 
 #[tauri::command]
-pub fn set_binding(state: State<'_, AppState>, key: LogicalKey, button: u16) {
+pub fn set_binding(state: State<'_, AppState>, key: LogicalKey, button: u32) {
     let _ = state.command_tx.send(MonitorCommand::SetKeyBinding { key, button });
 }
